+++ NEWS.md
@@ -1,6 +1,9 @@
 
 0.0.4
 ===================
+- Enable alternative overpass API services through `get_overpass_url()` and
+  `set_overpass_url()` functions
+- Extend and improve vignette
 
 0.0.3
 ===================